v4l2src: Support num-alloc-buffer option
This option uses to decide number of buffers will allocate.
And it only supports for io-mode=mmap/dmabuf
Default is auto decision, it will be calculated based on requirement
from driver, proposal from downstream element and so on.
If the value of num-alloc-buffer is too small or too big, number
of buffers will be adjusted to suit with supported range of driver.

Signed-off-by: Thao Luong <thao.luong.xh@renesas.com>

Porting from 1.18.5 to 1.22.11
Signed-off-by: duytruong <duy.truong.yw@renesas.com>